Clear metadata

+++ launcher.py
@@ -197,13 +197,17 @@ class xnat_pic_gui(tk.Frame):
             self.btn_overwrite = tk.Checkbutton(self.conv_popup, text="Overwrite existing folders", variable=self.overwrite_flag,
                                 onvalue=1, offvalue=0, command=checkOverwrite)
             self.btn_overwrite.grid(row=3, column=1, sticky='W')
-            # self.params = {'results_flag': master.results_flag,
-            #                 'overwrite_flag': master.overwrite_flag
-            #                 }
 
         def prj_conversion(self, master):
 
             ############### Whole project conversion ################
+
+            # Disable the buttons
+            master.convert_btn['state'] = tk.DISABLED
+            master.info_btn['state'] = tk.DISABLED
+            master.upload_btn['state'] = tk.DISABLED
+            master.process_btn['state'] = tk.DISABLED
+
             # Ask for project directory
             self.folder_to_convert = filedialog.askdirectory(parent=master.root, initialdir=os.path.expanduser("~"), title="XNAT-PIC: Select project directory in Bruker ParaVision format")
             if not self.folder_to_convert:
@@ -213,14 +217,11 @@ class xnat_pic_gui(tk.Frame):
             master.root.deiconify()
             master.root.update()
             self.dst = self.folder_to_convert + '_dcm'
+
+            # Initialize converter class
+            converter = Bruker2DicomConverter(self.params)
             
             try:
-                # Disable the buttons
-                master.convert_btn['state'] = tk.DISABLED
-                master.info_btn['state'] = tk.DISABLED
-                master.upload_btn['state'] = tk.DISABLED
-                master.process_btn['state'] = tk.DISABLED
-
                 start_time = time.time()
 
                 # Start the progress bar
@@ -264,10 +265,8 @@ class xnat_pic_gui(tk.Frame):
                                 os.makedirs(current_dst)
 
                         # Perform DICOM conversion
-                        converter = Bruker2DicomConverter(current_folder, current_dst, self.params)
                         # converter.start_conversion()
-                        # converter.multi_processes_conversion()
-                        tp = threading.Thread(target=converter.multi_core_conversion, args=())
+                        tp = threading.Thread(target=converter.multi_core_conversion, args=(current_folder, current_dst, ))
                         tp.start()
                         while tp.is_alive() == True:
                             # progressbar.update_bar()
@@ -301,6 +300,13 @@ class xnat_pic_gui(tk.Frame):
         def sbj_conversion(self, master):
 
             ############### Single subject conversion ################
+
+            # Convert from bruker to DICOM and disable the buttons
+            master.convert_btn['state'] = tk.DISABLED
+            master.info_btn['state'] = tk.DISABLED
+            master.upload_btn['state'] = tk.DISABLED
+            master.process_btn['state'] = tk.DISABLED
+
             # Ask for subject directory
             self.folder_to_convert = filedialog.askdirectory(parent=master.root, initialdir=os.path.expanduser("~"), title="XNAT-PIC: Select subject directory in Bruker ParaVision format")
             if not self.folder_to_convert:
@@ -313,6 +319,9 @@ class xnat_pic_gui(tk.Frame):
             head = head + '_dcm'
             project_foldername = tail.split('.',1)[0] + "_dcm"
             self.dst = os.path.join(head, project_foldername).replace('\\', '/')
+
+            # Start converter
+            converter = Bruker2DicomConverter(self.params)
 
             # If the destination folder already exists throw exception, otherwise create the new folder
             if os.path.isdir(self.dst):
@@ -336,22 +345,14 @@ class xnat_pic_gui(tk.Frame):
                     os.makedirs(self.dst)
 
             try:
-                # Convert from bruker to DICOM and disable the buttons
-                master.convert_btn['state'] = tk.DISABLED
-                master.info_btn['state'] = tk.DISABLED
-                master.upload_btn['state'] = tk.DISABLED
-                master.process_btn['state'] = tk.DISABLED
-
                 start_time = time.time()
 
                 # Start the progress bar
                 progressbar = ProgressBar(bar_title='DICOM Subject Converter')
                 progressbar.start_indeterminate_bar()
 
-                # Start converter
-                converter = Bruker2DicomConverter(self.folder_to_convert, self.dst, self.params)
                 # Initialize conversion thread
-                tp = threading.Thread(target=converter.multi_core_conversion, args=())
+                tp = threading.Thread(target=converter.multi_core_conversion, args=(self.folder_to_convert, self.dst, ))
                 tp.start()
                 while tp.is_alive() == True:
                     # As long as the thread is working, update the progress bar
@@ -447,8 +448,6 @@ class xnat_pic_gui(tk.Frame):
                 master.my_canvas.create_window(x_folder_list, y_folder_list1, width = int(my_width*25/100), height = int(my_height*40/100) ,anchor = tk.NW, window = my_listbox)
 
                 # List of subject in the project in the listbox
-                #results_prj = list(map(lambda i: results[i], range(1, len(results), len(fields))))
-                #my_listbox.insert(tk.END, *results_prj)
                 my_listbox.insert(tk.END, *item_list)
 
                 # Attach listbox to x and y scrollbar ()
@@ -493,17 +492,17 @@ class xnat_pic_gui(tk.Frame):
                     count += 1
 
                 # Menu
-                # Dose
+                # Group
                 OPTIONS = ["untreated", "treated"]
                 selected_group = tk.StringVar()
                 group_menu = ttk.Combobox(master.my_canvas, textvariable=selected_group, font = SMALL_FONT)
                 group_menu['values'] = OPTIONS
                 group_menu['state'] = 'disabled'
-                x_dose_menu = int(my_width*72/100)
-                y_dose_perc = 25
-                y_dose_delta = 7.5
-                y_dose_menu = int(my_height*(y_dose_perc+3*y_dose_delta)/100)
-                master.my_canvas.create_window(x_dose_menu, y_dose_menu, anchor = tk.NW, width = int(my_width*9.5/100), window = group_menu)
+                x_group_menu = int(my_width*72/100)
+                y_group_perc = 25
+                y_group_delta = 7.5
+                y_group_menu = int(my_height*(y_group_perc+3*y_group_delta)/100)
+                master.my_canvas.create_window(x_group_menu, y_group_menu, anchor = tk.NW, width = int(my_width*9.5/100), window = group_menu)
                 
                 # Timepoint
                 OPTIONS = ["pre", "post"]
@@ -512,7 +511,7 @@ class xnat_pic_gui(tk.Frame):
                 timepoint_menu['values'] = OPTIONS
                 timepoint_menu['state'] = 'disabled'
                 x_timepoint_menu = int(my_width*72/100)
-                y_timepiont_menu = int(my_height*(y_dose_perc+5*y_dose_delta)/100)
+                y_timepiont_menu = int(my_height*(y_group_perc+5*y_group_delta)/100)
                 master.my_canvas.create_window(x_timepoint_menu, y_timepiont_menu, anchor = tk.NW, width = int(my_width*5/100), window = timepoint_menu)
 
                 time_entry = tk.Entry(master.my_canvas, state='disabled', font = SMALL_FONT)
@@ -547,8 +546,29 @@ class xnat_pic_gui(tk.Frame):
                         entries[i].insert(0, str(results[selected_index*max_lim+i]))
                         entries[i].config(state=tk.DISABLED)
 
-                my_listbox.bind("<Double-Button-1>", items_selected)
+                my_listbox.bind('<Tab>', items_selected)
                 
+                #################### Clear the metadata ####################
+                clear_text = tk.StringVar() 
+                clear_btn = tk.Button(master.my_canvas, textvariable=clear_text, font=SMALL_FONT, bg=BG_BTN_COLOR, fg=TEXT_BTN_COLOR, borderwidth=0, command = lambda: clear_metadata(), cursor=CURSOR_HAND)
+                clear_text.set("Clear")
+                x_clear_btn = int(my_width*72/100)
+                y_clear_btn = int(my_height*(y_group_perc+1*y_group_delta)/100)
+                width_clear_btn = int(my_width*8/100)
+                master.my_canvas.create_window(x_clear_btn, y_clear_btn, anchor = tk.NW, width = width_clear_btn, window = clear_btn)
+                
+                def clear_metadata():
+                    # Clear all the combobox and the entry
+                    selected_group.set('')
+                    selected_timepoint.set('')
+                    selected_timepoint1.set('')
+                    time_entry.delete(0, tk.END)
+                    state = entries[0]['state']
+                    # Set empty string in all the entries
+                    for i in range(2, len(fields)):
+                            entries[i]['state'] = tk.NORMAL
+                            entries[i].delete(0, tk.END)
+                            entries[i]['state'] = state
                 #################### Modify the metadata ####################
                 modify_text = tk.StringVar() 
                 modify_btn = tk.Button(master.my_canvas, textvariable=modify_text, font=LARGE_FONT, bg=BG_BTN_COLOR, fg=TEXT_BTN_COLOR, borderwidth=0, command = lambda: modify_metadata(), cursor=CURSOR_HAND)
@@ -603,7 +623,7 @@ class xnat_pic_gui(tk.Frame):
                         selected_index
                         pass
                     except Exception as e:
-                         messagebox.showerror("XNAT-PIC", "Select a folder from the list box on the left")
+                         messagebox.showerror("XNAT-PIC", "Double click to select a folder from the list box on the left")
                          raise 
 
                     if not entries[fields.index("Project")].get():
@@ -679,36 +699,60 @@ class xnat_pic_gui(tk.Frame):
                 x_multiple_conf_btn = int(my_width*78/100)
                 master.my_canvas.create_window(x_multiple_conf_btn, y_btn, anchor = tk.NW, width = width_btn, window = multiple_confirm_btn)
                 
+                def normal_button():
+                    clear_btn["state"] = tk.NORMAL
+                    modify_btn["state"] = tk.NORMAL
+                    confirm_btn["state"] = tk.NORMAL
+                    multiple_confirm_btn["state"] = tk.NORMAL
+                    save_btn["state"] = tk.NORMAL
+
                 def confirm_multiple_metadata():
-
-                    messagebox.showinfo("Metadata","Select the folders from the box on the left for which to copy the info entered! Press ENTER to confirm or ESC to cancel!")
-                    
-                    confirm_metadata()
-
+                    clear_btn["state"] = tk.DISABLED
+                    modify_btn["state"] = tk.DISABLED
+                    confirm_btn["state"] = tk.DISABLED
+                    multiple_confirm_btn["state"] = tk.DISABLED
+                    save_btn["state"] = tk.DISABLED
+ 
+                    messagebox.showinfo("Metadata","1. Select the folders from the box on the left for which to copy the info entered!\n 2. Always remaining in the box on the left, press ENTER to confirm or ESC to cancel!")
+                    try:
+                        selected_index
+                        pass
+                    except Exception as e:
+                         normal_button()
+                         messagebox.showerror("XNAT-PIC", "Double click to select a folder from the list box on the left")
+                         raise 
                     my_listbox.selection_set(selected_index)    
                     my_listbox['selectmode'] = MULTIPLE
                     
                     # The user presses 'enter' to confirm 
                     def items_selected2(event):
-                       # Get indexes of selected folders
-                       selected_text_list = my_listbox.curselection()
                        
-                       # Update the list of results
-                       max_lim = len(fields)
-                       for i in range(0, len(selected_text_list)):
-                            for j in range(0, max_lim):
-                               results[selected_text_list[i]*max_lim+j] =  entries[j].get()
 
-                       # Update the txt file
-                       for i in range(0, len(selected_text_list)):
-                            with open(path_list[selected_text_list[i]], 'w+') as meta_file:
-                                                meta_file.write(tabulate([['Project', str(results[selected_index*max_lim+0])], ['Subject', str(results[selected_index*max_lim+1])], ['Acquisition_date', str(results[selected_index*max_lim+2])], 
-                       
-                                                ['Group', str(results[selected_index*max_lim+3])], ['Dose', str(results[selected_index*max_lim+4])], ['Timepoint', str(results[selected_index*max_lim+5])]], headers=['Variable', 'Value']))
+
+                       seltext = [my_listbox.get(index) for index in my_listbox.curselection()]
+                       result = messagebox.askquestion("Multiple Confirm", "Are you sure you want to save data for the following folders?\n" + '\n'.join(seltext), icon='warning')
+                       if result == 'yes':
+                        confirm_metadata()
+                        # Get indexes of selected folders
+                        selected_text_list = my_listbox.curselection()
+                        
+                        # Update the list of results
+                        max_lim = len(fields)
+                        for i in range(0, len(selected_text_list)):
+                                for j in range(0, max_lim):
+                                   results[selected_text_list[i]*max_lim+j] =  entries[j].get()
+
+                        # Update the txt file
+                        for i in range(0, len(selected_text_list)):
+                                with open(path_list[selected_text_list[i]], 'w+') as meta_file:
+                                                    meta_file.write(tabulate([['Project', str(results[selected_index*max_lim+0])], ['Subject', str(results[selected_index*max_lim+1])], ['Acquisition_date', str(results[selected_index*max_lim+2])], 
+                        
+                                                    ['Group', str(results[selected_index*max_lim+3])], ['Dose', str(results[selected_index*max_lim+4])], ['Timepoint', str(results[selected_index*max_lim+5])]], headers=['Variable', 'Value']))
                       
-                       messagebox.showinfo("Metadata","The information has been saved for the selected folders!")
+                        messagebox.showinfo("Metadata","The information has been saved for the selected folders!")
 
                        # Clear the focus and the select mode of the listbox is single
+                       normal_button()
                        my_listbox.selection_clear(0, 'end')
                        my_listbox['selectmode'] = SINGLE
                        
@@ -719,6 +763,7 @@ class xnat_pic_gui(tk.Frame):
                     def items_cancel(event):
                          # Clear the focus and the select mode of the listbox is single
                         messagebox.showinfo("Metadata","The information was not saved for the selected folders!")
+                        normal_button()
                         my_listbox.selection_clear(0, 'end')
                         my_listbox['selectmode'] = SINGLE
                     my_listbox.bind("<Escape>", items_cancel)
@@ -785,6 +830,7 @@ class xnat_pic_gui(tk.Frame):
                     exit_btn.destroy()
                     multiple_confirm_btn.destroy()
                     xnat_pic_gui.choose_you_action(master)
+
     # Upload files         
     class xnat_dcm_uploader():
